+++ eea/ldapadmin/users_admin.py
@@ -579,11 +579,6 @@ class UsersAdmin(SimpleItem, PropertyManager):
             agent = self._get_ldap_agent(bind=True)
             old_info = agent.user_info(user_id)
 
-            new_info.update(first_name=old_info['first_name'])
-            # allow changing the last name
-            # new_info.update(first_name=old_info['first_name'],
-            #                 last_name=old_info['last_name'])
-
             new_org_id = new_info['organisation']
             new_org_id_valid = agent.org_exists(new_org_id)